+++ go/libraries/doltcore/sqle/index/index_lookup.go
@@ -16,7 +16,10 @@ package index
 
 import (
 	"context"
+	"encoding/binary"
 	"fmt"
+	"io"
+	"sync"
 
 	"github.com/dolthub/go-mysql-server/sql"
 
@@ -29,13 +32,12 @@ import (
 func RowIterForIndexLookup(ctx *sql.Context, ilu sql.IndexLookup, columns []string) (sql.RowIter, error) {
 	lookup := ilu.(*doltIndexLookup)
 	idx := lookup.idx
-	rows := durable.NomsMapFromIndex(lookup.IndexRowData())
 
-	return RowIterForRanges(ctx, idx, lookup.ranges, rows, columns)
+	return RowIterForRanges(ctx, idx, lookup.ranges, lookup.IndexRowData(), columns)
 }
 
-func RowIterForRanges(ctx *sql.Context, idx DoltIndex, ranges []*noms.ReadRange, rowData types.Map, columns []string) (sql.RowIter, error) {
-	nrr := noms.NewNomsRangeReader(idx.IndexSchema(), rowData, ranges)
+func RowIterForRanges(ctx *sql.Context, idx DoltIndex, ranges []*noms.ReadRange, rowData durable.Index, columns []string) (sql.RowIter, error) {
+	nrr := noms.NewNomsRangeReader(idx.IndexSchema(), durable.NomsMapFromIndex(rowData), ranges)
 
 	covers := indexCoversCols(idx, columns)
 	if covers {
@@ -69,6 +71,61 @@ type IndexLookupKeyIterator interface {
 
 func DoltIndexFromLookup(lookup sql.IndexLookup) DoltIndex {
 	return lookup.(*doltIndexLookup).idx
+}
+
+func PartitionIndexedTableRows(ctx *sql.Context, idx sql.Index, projectedCols []string, part sql.Partition) (sql.RowIter, error) {
+	rp := part.(rangePartition)
+	ranges := []*noms.ReadRange{rp.partitionRange}
+	return RowIterForRanges(ctx, idx.(DoltIndex), ranges, rp.rowData, projectedCols)
+}
+
+func NewRangePartitionIter(lookup sql.IndexLookup) sql.PartitionIter {
+	dlu := lookup.(*doltIndexLookup)
+	return &rangePartitionIter{
+		ranges:  dlu.ranges,
+		curr:    0,
+		mu:      &sync.Mutex{},
+		rowData: dlu.IndexRowData(),
+	}
+}
+
+type rangePartitionIter struct {
+	ranges  []*noms.ReadRange
+	curr    int
+	mu      *sync.Mutex
+	rowData durable.Index
+}
+
+// Close is required by the sql.PartitionIter interface. Does nothing.
+func (itr *rangePartitionIter) Close(*sql.Context) error {
+	return nil
+}
+
+// Next returns the next partition if there is one, or io.EOF if there isn't.
+func (itr *rangePartitionIter) Next(_ *sql.Context) (sql.Partition, error) {
+	itr.mu.Lock()
+	defer itr.mu.Unlock()
+
+	if itr.curr >= len(itr.ranges) {
+		return nil, io.EOF
+	}
+
+	var bytes [4]byte
+	binary.BigEndian.PutUint32(bytes[:], uint32(itr.curr))
+	part := rangePartition{itr.ranges[itr.curr], bytes[:], itr.rowData}
+	itr.curr += 1
+
+	return part, nil
+}
+
+type rangePartition struct {
+	partitionRange *noms.ReadRange
+	keyBytes       []byte
+	rowData        durable.Index
+}
+
+func (rp rangePartition) Key() []byte {
+	return rp.keyBytes
 }
 
 type doltIndexLookup struct {
